pricing half way finished.  Minute Menu Plan cancelled.

+++ menuplanner/grails-app/domain/com/mp/domain/Quantity.groovy
@@ -6,7 +6,7 @@ import com.mp.tools.UnitUtil
 
 class Quantity {
 
-  Float value
+  BigDecimal value
   Unit unit
   Unit savedUnit
